Refoctor and enhance simulation flow
