+++ simulation/functions/all_vswarm_functions.yaml
@@ -26,7 +26,7 @@ services:
   ## Fibonacci ---------------------------------
   fibonacci-go:
     image: vhiveease/fibonacci-go:latest
-    container_name: fibonacci-go
+    container_name: function
     entrypoint:
       - /app/server
       - -addr=0.0.0.0:50051
@@ -36,7 +36,7 @@ services:
 
   fibonacci-nodejs:
     image: vhiveease/fibonacci-nodejs:latest
-    container_name: fibonacci-nodejs
+    container_name: function
     entrypoint:
       - node
       - server.js
@@ -48,7 +48,7 @@ services:
 
   fibonacci-python:
     image: vhiveease/fibonacci-python:latest
-    container_name: fibonacci-python
+    container_name: function
     entrypoint:
       - python
       - /app/server.py
@@ -62,7 +62,7 @@ services:
   ## AES ---------------------------------
   aes-go:
     image: vhiveease/aes-go:latest
-    container_name: aes-go
+    container_name: function
     entrypoint:
       - /app/server
       - -addr=0.0.0.0:50051
@@ -72,7 +72,7 @@ services:
 
   aes-nodejs:
     image: vhiveease/aes-nodejs:latest
-    container_name: aes-nodejs
+    container_name: function
     entrypoint:
       - node
       - server.js
@@ -84,7 +84,7 @@ services:
 
   aes-python:
     image: vhiveease/aes-python:latest
-    container_name: aes-python
+    container_name: function
     entrypoint:
       - python
       - /app/server.py
@@ -98,7 +98,7 @@ services:
   ## Authentication -----------------------
   auth-go:
     image: vhiveease/auth-go:latest
-    container_name: auth-go
+    container_name: function
     entrypoint:
       - /app/server
       - -addr=0.0.0.0:50051
@@ -108,7 +108,7 @@ services:
 
   auth-nodejs:
     image: vhiveease/auth-nodejs:latest
-    container_name: auth-nodejs
+    container_name: function
     entrypoint:
       - node
       - server.js
@@ -120,7 +120,7 @@ services:
 
   auth-python:
     image: vhiveease/auth-python:latest
-    container_name: auth-python
+    container_name: function
     entrypoint:
       - python
       - /app/server.py
